Refactored mid event for EditReturnedOrderController.java
