+++ src/main/java/uk/gov/hmcts/reform/prl/services/EditReturnedOrderService.java
@@ -39,8 +39,8 @@ public class EditReturnedOrderService {
 
     private static final String BOLD_BEGIN = "<span class='heading-h2'>";
     private static final String BOLD_END = "</span>";
-    private static final String INSTRUCTIONS_FROM_JUDGE = BOLD_BEGIN + "Instructions from the judge" + BOLD_END + " \n\n ";
-    private static final String EDIT_THE_ORDER_LABEL = "\n\n <span class='heading-h3'>Edit the order</span> \n\n";
+    private static final String INSTRUCTIONS_FROM_JUDGE = BOLD_BEGIN + "Instructions from the judge" + BOLD_END + " \n ";
+    private static final String EDIT_THE_ORDER_LABEL = "\n\n <span class='heading-h3'>Edit the order</span> \n";
     public static final String OPEN_THE_DRAFT_ORDER_TEXT = "Open the draft order and edit it to include the judge's instructions.";
     public static final String USE_CONTINUE_TO_EDIT_THE_ORDER = "Use continue to edit the order.";
     public static final String ORDER_UPLOADED_AS_DRAFT_FLAG = "orderUploadedAsDraftFlag";
@@ -186,5 +186,24 @@ public class EditReturnedOrderService {
             .isThisOrderFromHearing(caseData.getManageOrders().getIsThisOrderFromHearing())
             .selectTheHearing(caseData.getManageOrders().getSelectTheHearing())
             .build();
+    }
+
+    public AboutToStartOrSubmitCallbackResponse populateInstructionsAndFieldsForLegalRep(String authorisation, CallbackRequest callbackRequest) {
+        CaseData caseData = objectMapper.convertValue(
+            callbackRequest.getCaseDetails().getData(),
+            CaseData.class
+        );
+        if (caseData.getDraftOrderCollection() != null
+            && !caseData.getDraftOrderCollection().isEmpty()) {
+            DraftOrder selectedOrder = draftAnOrderService.getSelectedDraftOrderDetails(caseData.getDraftOrderCollection(),
+                                                                                        caseData.getManageOrders()
+                                                                                            .getRejectedOrdersDynamicList());
+            Map<String, Object> caseDataUpdated = populateInstructionsAndDocuments(caseData, authorisation, selectedOrder);
+            caseDataUpdated.putAll(draftAnOrderService.populateCommonDraftOrderFields(authorisation, caseData, selectedOrder));
+            return AboutToStartOrSubmitCallbackResponse.builder()
+                .data(caseDataUpdated).build();
+        } else {
+            return AboutToStartOrSubmitCallbackResponse.builder().errors(List.of("There are no draft orders")).build();
+        }
     }
 }
